Move encode/decode type key functions to openscreen-common
Closes #15
